fix(monitoring): irc_lurker resolve_active_session liest active_session_id als i64 (war i32)
Von #11s BIGINT-Fixture-Fix aufgedeckter Live-Bug: .ok() verschluckte den Decode-Fehler (bigint->i32), resolve_active_session lieferte in Prod immer None -> Chatter-Tracking (upsert_names_batch/upsert_chatter_seen) war No-op.

+++ rust/crates/tb-monitoring/src/irc_lurker.rs
@@ -92,8 +92,8 @@ fn is_irc_word(c: char) -> bool {
 // ---- DB-Layer --------------------------------------------------------------
 
 /// Aktive Session-ID eines live Kanals (`twitch_live_state`), sonst `None`.
-async fn resolve_active_session(pool: &PgPool, channel: &str) -> Option<i32> {
-    sqlx::query_scalar::<_, Option<i32>>(
+async fn resolve_active_session(pool: &PgPool, channel: &str) -> Option<i64> {
+    sqlx::query_scalar::<_, Option<i64>>(
         "SELECT active_session_id FROM twitch_live_state \
          WHERE LOWER(streamer_login) = $1 AND is_live = 1",
     )
